+++ src/OpenTelemetry/CHANGELOG.md
@@ -8,7 +8,7 @@ Notes](../../RELEASENOTES.md).
 
 * Fixed precision issues when using `Histogram<float>` with custom
   `HistogramBucketBoundaries`.
-  ([#6803](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6803))
+  ([#6866](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6866))
 * Fixed a thread-safety issue in `LogRecordSharedPool.Rent()`.
   ([#6833](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6833)