+++ src/OpenTelemetry/CHANGELOG.md
@@ -10,6 +10,9 @@ Notes](../../RELEASENOTES.md).
 produced `long.MinValue`.
 ([[#6928](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6928)])
 
+* Fixed precision issues when using `Histogram<float>` with custom
+  `HistogramBucketBoundaries`.
+  ([#6866](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6866))
 * Fixed a thread-safety issue in `LogRecordSharedPool.Rent()`.
   ([#6833](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6833)