+++ src/OpenTelemetry/CHANGELOG.md
@@ -10,6 +10,13 @@ Notes](../../RELEASENOTES.md).
   the previous activity was stopped by another thread during span creation.
   ([#6257](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6257))
 
+* Fixed `OverflowException` in `TraceIdRatioBasedSampler` when trace ID bytes
+produced `long.MinValue`.
+([[#6928](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6928)])
+
+* Fixed precision issues when using `Histogram<float>` with custom
+  `HistogramBucketBoundaries`.
+  ([#6866](https://github.com/open-telemetry/opentelemetry-dotnet/issues/6866))
 * Fixed a thread-safety issue in `LogRecordSharedPool.Rent()`.
   ([#6833](https://github.com/open-telemetry/opentelemetry-dotnet/pull/6833))